new version unit product configuration
